requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
Django==3.2.15
djangorestframework==3.13.1
idna==3.4
Pillow==9.2.0
psycopg2==2.9.3
pycparser==2.21
python-dotenv==0.21.0
pytz==2022.4
requests==2.28.1
sqlparse==0.4.3
typing_extensions==4.3.0
urllib3==1.26.12
